requirements.txt
aiosqlite==0.19.0
alembic==1.11.1
altair==5.0.1
ansi2html==1.8.0
anyio==3.7.0
apprise==1.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==2.1.0
astroid==2.15.5
asttokens==2.2.1
asyncpg==0.27.0
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
colour==0.1.5
comm==0.1.3
contourpy==1.1.0
coolname==2.2.0
croniter==1.3.15
cryptography==41.0.1
cycler==0.11.0
dacite==1.8.1
dash==2.10.2
dash-auth==2.0.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.7
dateparser==1.1.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.7
docker==6.1.3
dtreeviz==2.2.1
entrypoints==0.4
evidently==0.3.3
exceptiongroup==1.1.1
executing==1.2.0
explainerdashboard==0.4.2.2
fastapi==0.97.0
fastjsonschema==2.17.1
filelock==3.12.2
Flask==2.3.2
flask-simplelogin==0.1.2
Flask-WTF==1.1.1
fonttools==4.40.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
greenlet==2.0.2
griffe==0.29.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmlmin==0.1.12
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
hyperframe==6.0.1
identify==2.5.26
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.23.2
ipython==8.12.2
ipython-genutils==0.2.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.1
jupyter-dash==0.4.2
jupyter-events==0.6.3
jupyter-server==2.6.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
kubernetes==26.1.0
lazy-object-proxy==1.9.0
llvmlite==0.40.1
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
mlflow==2.4.1
multimethod==1.9.1
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.4
notebook-shim==0.2.3
numba==0.57.1
numpy==1.23.0
oauthlib==3.2.2
orjson==3.9.1
overrides==7.3.1
oyaml==1.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
patsy==0.5.3
pendulum==2.1.2
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
pip==20.0.2
pkg-resources==0.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
plotly==5.15.0
pluggy==1.2.0
pre-commit==3.3.3
prefect==2.82
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.17.4
Pympler==1.0.1
pynndescent==0.5.10
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.3.2
pytest==7.3.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
querystring-parser==1.2.4
readchar==4.0.5
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==44.0.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
streamlit==1.24.0
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typeguard==2.13.3
typer==0.9.0
typing-extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
umap-learn==0.5.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
validators==0.20.0
virtualenv==20.24.2
visions==0.7.5
waitress==2.1.2
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.3
websockets==11.0.3
Werkzeug==2.2.3
wheel==0.34.2
wordcloud==1.9.2
wrapt==1.15.0
WTForms==3.0.1
xgboost==1.7.5
ydata-profiling==4.3.0
zipp==3.15.0
